backend/requirements.txt
Django==5.1.1
psycopg2-binary==2.9.9
Pillow==10.4.0

djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1

django-environ==0.11.2
boto3==1.34.162
django-storages==1.14.4
minio==7.2.7

django-debug-toolbar==4.4.6
